openclaw_mt5_bridge/requirements.txt
fastapi
uvicorn
MetaTrader5
python-dotenv
pydantic
pytest
pandas
